requirements.txt
requests
PyYAML>=5.1
python-dotenv>=0.21.0